Fix little signed/unsigned integer comparison warning

+++ rng/unix/mc_getrandom_stubs.c
@@ -35,8 +35,8 @@ void raw_getrandom (uint8_t *data, uint32_t len) {
 #include <sys/param.h>
 
 void raw_getrandom (uint8_t *data, uint32_t len) {
-  int rlen = 0;
-  for (int i = 0; i <= len; i += 256) {
+  size_t rlen = 0;
+  for (uint32_t i = 0; i <= len; i += 256) {
     rlen = MIN(256, len - i);
     if (getentropy(data + i, rlen) < 0) uerror("getentropy", Nothing);
   }
